Phase05_AdvancedCloudDeployment/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "modelcontextprotocol>=1.0.1",
    "openai-agents>=0.6.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "starlette>=0.50.0",
    "uvicorn[standard]>=0.40.0",
]
